+++ changelog.md
@@ -10,7 +10,7 @@ NeoForge, and Fabric
 
 ### Changed
 
-- Match FD 1.2.10 - 1.21.1 Tag Changes:
+- Match FD 1.2.10 - 1.21.1+ Tag Changes:
   - All usages of `c:foods/milk` -> `c:drinks/milk`
   - Pie/Cheesecake added to `c:foods/edible_when_placed`
     - [1.21.1 only] Punch Bowls added to `c:foods/edible_when_placed`
@@ -18,7 +18,7 @@ NeoForge, and Fabric
   - All Punch added to `c:drinks/punch`
   - All Syrup added to `c:drinks/syrup`
   - Added `c:drinks/punch` & `c:drinks/syrup` to `c:drinks`
-  - [1.21.1 only] Removed syrups from `c:foods`
+- [1.20.1, 1.21.1] Removed syrups from `c:foods`
 - Added Syrups, Monster Mash, and Slimeball Spaghetti to `create:upright_on_belt`
 - MidnightLib version updated